pyproject.toml
[build-system]
requires = [
    "hatchling>=1.19.0",
]
build-backend = "hatchling.build"

[project]
name = "mkdocs-yamp"
description = "Yet Another Multirepo Plugin for MkDocs. Adds additional repositories to a directory within the docs directory to be referenced in the navigation"
dynamic = ["version"]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=2.7"
authors = [
    { name = "Steven Terrana", email = "steventerrana@gmail.com" },
]
keywords = [
    "mkdocs",
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.4",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
]
dependencies = [
    "mkdocs>=1.5.3",
    "GitPython>=3.1.27"
]

[project.entry-points."mkdocs.plugins"]
my_mkdocs_plugin = "my_mkdocs_plugin.my_mkdocs_plugin:MyCustomPlugin"
editurl = "mkdocs_editurl_plugin.plugin:EditUrlPlugin"

[project.urls]
Homepage = "https://github.com/boozallen/mkdocs-yamp-plugin"

[tool.hatch.version]
path = "my_mkdocs_plugin/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/my_mkdocs_plugin", "/mkdocs_editurl_plugin"
]
exclude = [
  '/.github',
  '/docs'
]

[tool.hatch.build.targets.wheel]
include = [
    "/my_mkdocs_plugin", "/mkdocs_editurl_plugin"
]
exclude = [
  '/.github',
  '/docs'
]
